requirements.txt
greenlet==2.0.1
inflect==6.0.2
psycopg2-binary==2.9.5
pydantic==1.10.2
sqlacodegen==2.3.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.45
Flask==2.2.2
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==0.19.2
structlog==22.1.0
GeoAlchemy2==0.12.5